Move workshop list table mods to admin.php file
