+++ wp-content/plugins/wporg-learn/inc/post-type.php
@@ -10,8 +10,6 @@ defined( 'WPINC' ) || die();
  * Actions and filters.
  */
 add_action( 'init', __NAMESPACE__ . '\register' );
-add_filter( 'manage_wporg_workshop_posts_columns', __NAMESPACE__ . '\add_workshop_list_table_columns' );
-add_action( 'manage_wporg_workshop_posts_custom_column', __NAMESPACE__ . '\render_workshop_list_table_columns', 10, 2 );
 add_filter( 'jetpack_copy_post_post_types', __NAMESPACE__ . '\jetpack_copy_post_post_types' );
 add_filter( 'jetpack_sitemap_post_types', __NAMESPACE__ . '\jetpack_sitemap_post_types' );
 add_filter( 'jetpack_page_sitemap_other_urls', __NAMESPACE__ . '\jetpack_page_sitemap_other_urls' );
@@ -194,53 +192,6 @@ function generate_workshop_template_structure() {
 	);
 
 	return $template;
-}
-
-/**
- * Add additional columns to the post list table for workshops.
- *
- * @param array $columns
- *
- * @return array
- */
-function add_workshop_list_table_columns( $columns ) {
-	$columns = array_slice( $columns, 0, -2, true )
-				+ array( 'video_language' => __( 'Language', 'wporg-learn' ) )
-				+ array( 'video_caption_language' => __( 'Captions', 'wporg-learn' ) )
-				+ array_slice( $columns, -2, 2, true );
-
-	return $columns;
-}
-
-/**
- * Render the cell contents for the additional columns in the post list table for workshops.
- *
- * @param string $column_name
- * @param int    $post_id
- *
- * @return void
- */
-function render_workshop_list_table_columns( $column_name, $post_id ) {
-	$post = get_post( $post_id );
-
-	switch ( $column_name ) {
-		case 'video_language':
-			echo esc_html( get_locale_name_from_code( $post->video_language, 'english' ) );
-			break;
-		case 'video_caption_language':
-			$captions = get_post_meta( $post->ID, 'video_caption_language' );
-
-			echo esc_html( implode(
-				', ',
-				array_map(
-					function( $caption_lang ) {
-						return get_locale_name_from_code( $caption_lang, 'english' );
-					},
-					$captions
-				)
-			) );
-			break;
-	}
 }
 
 /**
